Revertir la acción de separación de 25ecb19 que ha llevado a separar el código de generación de las tarjetas de rango del módulo canvasProcessing.
