+++ modules/canvasProcessing.js
@@ -4,6 +4,7 @@ const randomstring = require('randomstring')
 const isValidUrl = require('is-valid-http-url')
 const isImageUrl = require('is-image-url')
 const hexToRgba = require('hex-to-rgba')
+const { millify } = require('millify')
 
 registerFont('./modules/sources/fonts/Montserrat/Montserrat-SemiBold.ttf', { family: 'Montserrat' })
 module.exports = {
@@ -65,6 +66,76 @@ module.exports = {
 
     const avatar = await loadImage(member.user.displayAvatarURL({ format: 'png', size: 512 }))
     ctx.drawImage(avatar, canvas.width / 2 - 100, 75, 200, 200)
+
+    const buffer = canvas.toBuffer('image/png')
+    writeFileSync(paths.attachmentSent, buffer)
+
+    return paths
+  },
+  generateRankCard: async (member, guildConfig) => {
+    const uniqueIdentifiers = {
+      userAvatar: randomstring.generate({ charset: 'alphabetic' }),
+      attachmentSent: randomstring.generate({ charset: 'alphabetic' })
+    }
+
+    const paths = {
+      attachmentSent: `./modules/temp/${uniqueIdentifiers.attachmentSent}.png`
+    }
+
+    const canvas = createCanvas(1100, 320)
+    const ctx = canvas.getContext('2d')
+
+    // Establecer fondo del canvas
+    let imgPath = ''
+    if (guildConfig.levelsImageCustomBackground && isValidUrl(guildConfig.levelsImageCustomBackground) && isImageUrl(guildConfig.levelsImageCustomBackground)) {
+      imgPath = guildConfig.levelsImageCustomBackground
+      const background = await loadImage(imgPath)
+      const scale = Math.max(canvas.width / background.width, canvas.height / background.height)
+      ctx.drawImage(background, (canvas.width / 2) - (background.width / 2) * scale, (canvas.height / 2) - (background.height / 2) * scale, background.width * scale, background.height * scale)
+
+      // Establecer blured overlay
+      ctx.fillStyle = hexToRgba(guildConfig.levelsImageCustomOverlayColor || '#272934', (guildConfig.levelsImageCustomOpacity / 100))
+      ctx.fillRect(25, 25, 1050, 270)
+    } else {
+      ctx.fillStyle = '#272934'
+      ctx.fillRect(0, 0, canvas.width, canvas.height)
+    }
+
+    // Escribir usuario
+    ctx.font = applyText(canvas, member.tag, 40)
+    ctx.textAlign = 'left'
+    ctx.fillStyle = 'rgba(255, 255, 255, 0.8)'
+    ctx.fillText(`${member.tag}`, 295, 180, 500)
+
+    // Escribir nivel, experiencia y rango
+
+    ctx.font = '50px "Montserrat SemiBold"'
+    ctx.fillStyle = 'rgba(255, 255, 255, 0.5)'
+    ctx.textAlign = 'right'
+    ctx.fillText(`Rank #${member.levelData.rank}  Level ${millify(member.levelData.memberLevel)}`, 1050, 100)
+
+    // Escribir progreso actual (actual/necesario)
+    const actualVSrequired = `${millify(member.levelData.memberExperience)} / ${millify(((member.levelData.memberLevel * member.levelData.memberLevel) * guildConfig.levelsDifficulty) * 100)} XP`
+
+    ctx.font = '30px "Montserrat SemiBold"'
+    ctx.textAlign = 'right'
+    ctx.fillStyle = 'rgba(255, 255, 255, 0.8)'
+    ctx.fillText(actualVSrequired, 1050, 180)
+
+    // Añadir barra de progreso (backdrop)
+
+    ctx.fillStyle = 'rgba(255,255,255, 0.3)'
+    ctx.fillRect(295, 200, 755, 70)
+
+    // Añadir barra de progreso
+
+    ctx.fillStyle = 'rgb(255,255,255)'
+    ctx.fillRect(295, 200, (Math.abs((member.levelData.memberExperience) / (((member.levelData.memberLevel * member.levelData.memberLevel) * guildConfig.levelsDifficulty) * 100)) * 755), 70)
+
+    // Añadir avatar de usuario
+
+    const avatar = await loadImage(member.user.displayAvatarURL({ format: 'png', size: 512 }))
+    ctx.drawImage(avatar, 50, 50, 220, 220)
 
     const buffer = canvas.toBuffer('image/png')
     writeFileSync(paths.attachmentSent, buffer)
